+++ src/MillisChronoTimer.cpp
@@ -20,11 +20,19 @@ bool MillisChronoTimer::expired() const {
     else return false;
 }
 
+double MillisChronoTimer::elapsed() const {
+    return elapsedSec() * 1000;  // Convert to milliseconds
+}
+
 double MillisChronoTimer::elapsedSec() const {
     auto now = std::chrono::high_resolution_clock::now();
     return std::chrono::duration<double>(now - start).count();  // duration in seconds
 }
 
-double MillisChronoTimer::elapsed() const {
-    return elapsedSec() * 1000;  // Convert to milliseconds
+double MillisChronoTimer::elapsedMin() const {
+    return elapsedSec() / 60.0;  // Convert to minutes
+}
+
+double MillisChronoTimer::elapsedHrs() const {
+    return elapsedSec() / 3600.0;  // Convert to hours
 }