+++ docs/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log MillisChronoTimer
 
+## [1.2.0] - 2025-02-20
+- Added methods to return elapsed: Minutes and Hours.
+- Added example of ms_sec_min_hrs.cpp
+
 ## [1.0.1] - 2025-02-07
 - File reorganization